requirements.txt
openai
requests
pillow
pyyaml
matplotlib
tqdm
colorlog
datasets>=2.15.0  # For HuggingFace dataset support
